+++ src/main/java/org/sonar/plugins/delphi/pmd/rules/UnusedImportsRule.java
@@ -9,4 +9,9 @@ public class UnusedImportsRule extends AbstractImportRule {
     UnitNameDeclaration dependency = unitImport.getImportNameDeclaration().getOriginalDeclaration();
     return !getUnitDeclaration().hasDependency(dependency);
   }
+
+  @Override
+  public String getMessage() {
+    return "Review this potentially unnecessary import";
+  }
 }